chore(pr) Moved the waitUntil function.
