+++ editor/src/core/shared/promise-utils.ts
@@ -9,3 +9,17 @@ export function cachedPromise<T>(promiseId: string, constructor: () => Promise<T
     return newPromise
   }
 }
+
+export async function waitUntil(limitMS: number, check: () => boolean): Promise<boolean> {
+  if (check()) {
+    return true
+  } else if (limitMS <= 0) {
+    return false
+  } else {
+    return new Promise((resolve) => {
+      setTimeout(resolve, 5)
+    }).then(() => {
+      return waitUntil(limitMS - 5, check)
+    })
+  }
+}
